+++ wandb/data_types.py
@@ -13,7 +13,6 @@ import binascii
 import codecs
 import json
 import logging
-import numbers
 import os
 import pprint
 import sys
@@ -24,7 +23,6 @@ from six.moves.collections_abc import Sequence
 import wandb
 from wandb import util
 from wandb.compat import tempfile
-from wandb.util import has_num
 
 _PY3 = sys.version_info.major == 3 and sys.version_info.minor >= 6
 
@@ -39,8 +37,8 @@ if _PY3:
         Molecule,
         Html,
         Video,
-        JSONMetadata,
         ImageMask,
+        BoundingBoxes2D,
     )
 else:
     from wandb.sdk_py27.interface import _dtypes
@@ -53,8 +51,8 @@ else:
         Molecule,
         Html,
         Video,
-        JSONMetadata,
         ImageMask,
+        BoundingBoxes2D,
     )
 
 __all__ = [
@@ -67,6 +65,7 @@ __all__ = [
     "Html",
     "Video",
     "ImageMask",
+    "BoundingBoxes2D",
 ]
 
 
@@ -1170,147 +1169,6 @@ class Image(BatchableMedia):
             and self._image == other._image
             and self._classes == other._classes
         )
-
-
-class BoundingBoxes2D(JSONMetadata):
-    """
-    Wandb class for 2D bounding boxes
-    """
-
-    artifact_type = "bounding-boxes"
-
-    def __init__(self, val, key, **kwargs):
-        """
-        Args:
-            val (dict): dictionary following the form:
-            {
-                "class_labels": optional mapping from class ids to strings {id: str}
-                "box_data": list of boxes: [
-                    {
-                        "position": {
-                            "minX": float,
-                            "maxX": float,
-                            "minY": float,
-                            "maxY": float,
-                        },
-                        "class_id": 1,
-                        "box_caption": optional str
-                        "scores": optional dict of scores
-                    },
-                    ...
-                ],
-            }
-            key (str): id for set of bounding boxes
-        """
-        super(BoundingBoxes2D, self).__init__(val)
-        self._val = val["box_data"]
-        self._key = key
-        # Add default class mapping
-        if "class_labels" not in val:
-            np = util.get_module(
-                "numpy", required="Semantic Segmentation mask support requires numpy"
-            )
-            classes = (
-                np.unique(list(map(lambda box: box["class_id"], val["box_data"])))
-                .astype(np.int32)
-                .tolist()
-            )
-            class_labels = dict((c, "class_" + str(c)) for c in classes)
-            self._class_labels = class_labels
-        else:
-            self._class_labels = val["class_labels"]
-
-    def bind_to_run(self, run, key, step, id_=None):
-        # bind_to_run key argument is the Image parent key
-        # the self._key value is the mask's sub key
-        super(BoundingBoxes2D, self).bind_to_run(run, key, step, id_=id_)
-        run._add_singleton(
-            "bounding_box/class_labels",
-            key + "_wandb_delimeter_" + self._key,
-            self._class_labels,
-        )
-
-    def type_name(self):
-        return "boxes2D"
-
-    def validate(self, val):
-        # Optional argument
-        if "class_labels" in val:
-            for k, v in list(val["class_labels"].items()):
-                if (not isinstance(k, numbers.Number)) or (
-                    not isinstance(v, six.string_types)
-                ):
-                    raise TypeError(
-                        "Class labels must be a dictionary of numbers to string"
-                    )
-
-        boxes = val["box_data"]
-        if not isinstance(boxes, Sequence):
-            raise TypeError("Boxes must be a list")
-
-        for box in boxes:
-            # Required arguments
-            error_str = "Each box must contain a position with: middle, width, and height or \
-                    \nminX, maxX, minY, maxY."
-            if "position" not in box:
-                raise TypeError(error_str)
-            else:
-                valid = False
-                if (
-                    "middle" in box["position"]
-                    and len(box["position"]["middle"]) == 2
-                    and has_num(box["position"], "width")
-                    and has_num(box["position"], "height")
-                ):
-                    valid = True
-                elif (
-                    has_num(box["position"], "minX")
-                    and has_num(box["position"], "maxX")
-                    and has_num(box["position"], "minY")
-                    and has_num(box["position"], "maxY")
-                ):
-                    valid = True
-
-                if not valid:
-                    raise TypeError(error_str)
-
-            # Optional arguments
-            if ("scores" in box) and not isinstance(box["scores"], dict):
-                raise TypeError("Box scores must be a dictionary")
-            elif "scores" in box:
-                for k, v in list(box["scores"].items()):
-                    if not isinstance(k, six.string_types):
-                        raise TypeError("A score key must be a string")
-                    if not isinstance(v, numbers.Number):
-                        raise TypeError("A score value must be a number")
-
-            if ("class_id" in box) and not isinstance(
-                box["class_id"], six.integer_types
-            ):
-                raise TypeError("A box's class_id must be an integer")
-
-            # Optional
-            if ("box_caption" in box) and not isinstance(
-                box["box_caption"], six.string_types
-            ):
-                raise TypeError("A box's caption must be a string")
-
-    def to_json(self, run_or_artifact):
-        wandb_run, wandb_artifacts = _safe_sdk_import()
-
-        if isinstance(run_or_artifact, wandb_run.Run):
-            return super(BoundingBoxes2D, self).to_json(run_or_artifact)
-        elif isinstance(run_or_artifact, wandb_artifacts.Artifact):
-            # TODO (tim): I would like to log out a proper dictionary representing this object, but don't
-            # want to mess with the visualizations that are currently available in the UI. This really should output
-            # an object with a _type key. Will need to push this change to the UI first to ensure backwards compat
-            return self._val
-        else:
-            raise ValueError("to_json accepts wandb_run.Run or wandb_artifact.Artifact")
-
-    @classmethod
-    def from_json(cls, json_obj, source_artifact):
-        return cls({"box_data": json_obj}, "")
 
 
 class Plotly(Media):
